Refactor Acl, AclOption into acl module.
